Add the ability to declare columns without a corresponding field
This is needed for those cases where the importer will be using those
columns in non-conventional ways, like when used inside the .hook method
for doing more complex computations with that column's value.  In those
cases, the importer still needs to declare the column so that the
spreadsheet is checked to ensure its presence.